발생하는 쿼리 확인 후 쿼리 성능 개선 (#758)
* refactor: (#731) 커스텀 예외로 수정

* refactor: (#731) 정렬 타입을 created_at -> id로 변경

* refactor: (#731) 댓글 수 @Formula 및 서브 쿼리 제거

* refactor: (#731) 게시글 목록 조회 시 vote_count를 통한 정렬

* refactor: (#731) 멤버의 게시글 개수와 투표 개수 설정

* test: (#731) 수정된 기능 테스트 추가

* refactor: (#731) 랭킹 조회 기능 리팩토링

* fix: (#731) 잘못된 예외 처리 수정

* refactor: (#731) ID를 통해 댓글 목록 조회

* refactor: (#731) 개행 컨벤션 수정

* refactor: (#731) score 초기화

* refactor: (#731) 개행 컨벤션 수정

+++ backend/src/main/java/com/votogether/domain/member/dto/response/MemberInfoResponse.java
@@ -19,9 +19,9 @@ public record MemberInfoResponse(
         Roles roles,
 
         @Schema(description = "작성한 게시글 수", example = "5")
-        int postCount,
+        long postCount,
 
         @Schema(description = "투표한 수", example = "10")
-        int voteCount
+        long voteCount
 ) {
 }
